+++ common/flatpak-run-private.h
@@ -200,11 +200,4 @@ gboolean flatpak_run_app (FlatpakDecomposed  *app_ref,
 
 extern const char * const *flatpak_abs_usrmerged_dirs;
 
-gboolean flatpak_run_parse_x11_display (const char  *display,
-                                        int         *family,
-                                        char       **x11_socket,
-                                        char       **remote_host,
-                                        char       **original_display_nr,
-                                        GError     **error);
-
 #endif /* __FLATPAK_RUN_H__ */
