+++ common/meson.build
@@ -154,6 +154,11 @@ sources = [
   'flatpak-remote-ref.c',
   'flatpak-remote.c',
   'flatpak-run.c',
+  'flatpak-run-cups.c',
+  'flatpak-run-dbus.c',
+  'flatpak-run-pulseaudio.c',
+  'flatpak-run-sockets.c',
+  'flatpak-run-x11.c',
   'flatpak-transaction.c',
   'flatpak-utils-http.c',
   'flatpak-utils.c',
